dyslexia-lens-main/dyslexia-lens-main/ai-engine/requirements.txt
fastapi
uvicorn
python-multipart
google-generativeai
python-dotenv
pypdf
openai
anthropic
